Internal change
PiperOrigin-RevId: 342910130
